separate the optimism transaction type list from the mainnet transaction list
